Pull out general purpose LockedResources impl
